+++ tests/test_auxiliary.py
@@ -7,18 +7,19 @@
 # SPDX-License-Identifier: EPL-2.0
 ##########################################################################
 
-from multiprocessing.connection import deliver_challenge
-import time
 import logging
+import time
+from multiprocessing.connection import deliver_challenge
+
 import pytest
 
+import pykiso
 from pykiso import (
     AuxiliaryInterface,
     MpAuxiliaryInterface,
     SimpleAuxiliaryInterface,
     logging_initializer,
 )
-import pykiso
 from pykiso.auxiliary import AuxiliaryCommon
 from pykiso.exceptions import AuxiliaryCreationError
 from pykiso.test_setup.config_registry import ConfigRegistry
@@ -71,7 +72,7 @@ def mock_mp_aux(mocker):
     class MockMpAux(MpAuxiliaryInterface):
         def __init__(self, param_1=None, param_2=None, **kwargs):
             logging_initializer.log_options = logging_initializer.LogOptions(
-                None, "ERROR", None
+                None, "ERROR", None, False
             )
             self.param_1 = param_1
             self.param_2 = param_2